fleet_server/api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0   # JWT
bcrypt>=4.0.0                       # bcrypt хешування паролів
httpx>=0.27.0                       # HTTP клієнт (OAuth + WebSocket live poll)
python-dotenv>=1.0.0
jinja2>=3.1.0
websockets>=12.0
email-validator>=2.1.0              # Pydantic EmailStr
python-multipart>=0.0.9             # Form data
